Add private -[PMDatabase _database] method
This method is meant to be in the package scope. It allows other
classes in the same package to access the pmdb_t in the database
object. This allows us to remove a lot of unecessary code when
initializing a PMPackage, and simply use the reference to the pmdb_t
directly.

Signed-off-by: Sebastian Nowicki <sebnow@gmail.com>